requirements.txt
Flask==1.1.1
Jinja2==3.0.3
MarkupSafe==2.0
Click==8.0.0
psycopg2-binary==2.8.4
Werkzeug==0.15.5

